Cisco: update_port should only invoke n1kv and not nexus plugin
Update port method should not invoke the nexus plugin when it is
not configured with n1kv plugin. This patch adds a condition to the
update_port method making sure that nexus plugin is invoked only
when it is configured in cisco_plugins.ini file. This bug is only
targeted for the stable/icehouse release since the nexus monolithic
plugin code has been removed from Juno release onwards.

Change-Id: I67f023e8610975d9636124e9f82336b5ab407c16
Closes-bug: #1387846