Send tenant_id 0 for shared networks to VSM
We need to indicate if a network is shared for VSM. In order
to do so, we will now send tenant_id = '0' for shared networks. This
change additionally ensures the tenant_id is properly updated when the
network is updated to be a shared network after it is created.

Also removed broken test case in monolithic plugin UTs and fixed
copyright lines in UT files.

Closes-Bug: #1440214
Change-Id: I50c0d19a47f82994a9b17fc70ea9e2fba6deec39