Implement update port for N1kV ML2 mechanism driver
Added support for update port operation in order to have this operation
work successfully. On update port we check if this is the the first time
binding of a router or DHCP port as currently that is the only operation
we want to support updates for.

Closes-Bug: #1447369
Change-Id: I69624d7b0cb0eb82bc5cf6ed19202e076bee7cc4
(cherry picked from commit e7230174792199f932de92492e537fe263d18824)