Sanitize policy profile table on neutron restart
On neutron restart, VSM IPs can be reconfigured. We need to sanitize
the policy profile table and delete entries for VSMs that are not
configured. Delete will only take place if the profile is not in use
or if there is another VSM with the profile configured.

Change-Id: I631417da254d12184b71314391737173ab2e03c1
Closes-Bug: #1446892
(cherry picked from commit 2c1876be1690063b13c4a3aa5ad0e42144ca8009)