Call to get nexus type for Nexus 9372PX fails
Started seeing the error "Get call failed to return Nexus type" when
setting up network with 2 switches.  There was only one Nexus switch
in which error appeared on and it was the Nexus9372PX.  It seems
the XML output from EXEC_GET_INVENTORY_SNIPPET from both these
N9K were not consistent.  The word Chassis was capitalized in
9396 while not in 9372.

Change-Id: Ib0f44fe59229e0b2f945addcbcd624ed40e68340
Closes-bug:  #1466164