Replace retry count with replay failure stats
This code-set eliminates checking switch config replay attempts against
max configurable attempts.  Instead it will continue to contact the
switch forever.

I've repurposed some of the logic to provide statistics to
identify if there are issues in replay process.  There are two
statistics: 1) contact 2) config.  The contact stat identifies
a problem contacting the switch.  The operator would need to
see if there is something wrong with the nexus switch or the path
to the switch. The config stat identifies an issues with the config
set we're trying to send to the switch.  This latter case may require
part of the config to be removed.

Change-Id: Icb15d0b2d97f1c1c9f583989823a4dae256f3e65
Closes-bug:  #1469839
(cherry picked from commit e6673d534cdff2d5c8434ef9928a766b4304b003)