Make vNIC names explicit in UCSM plugin
This fix takes the ethenet port names of the ports that can be used to
carry traffic from Neutron virtual ports as user input in the UCSM
plugin config file. These ethernet port names have to match the names
on the UCS Manager. The plugin then uses this config to configure the
correct ethernet ports on the Service Profile for the server where the
virtual port is being created/instantiated.

Change-Id: I69a1dab1ec7e828ba36784c4b10b09762d8964fe
Closes-Bug: #1470556