Handle new baremetal format of port-id field
Ironic/Baremetal will learn the port id via LLDP.  As a result, the
format of this field can be different than what is received today.
Currently ML2 Nexus code expects ethernet:n/m or just n/m.  Baremetal
port-id format instead is Ethernet n/m or Ethernetn/m.  This port-id is
converted to all lowercase and whitespace is removed to provide
internal uniformity in this field and to minimize output variation in
unit tests.   During the split of internal interface name, the
interface type results in ethernet and port is n/m which is consistent
results with current ethernet formats.

Change-Id: Ifb9d6dbae00b510b95017e6fd0a1e4720b875ceb
Closes-bug:  #1544301