Catch DB exceptions defined in oslo_db instead of MySQL
DB exceptions in oslo_db are DB backend independent so they
should preferably be used. This patch therefore makes changes
so that oslo_db's DBReferenceError is catched. The patch
also moves an session expiry call out from the substransaction.
This is to reduce the risk that the above exception is
encountered when creating a binding between hosted port and
its hosting port.

Closes bug: #1551756

Change-Id: Ibc473bc5ded749bdc4154e7bd5f36eb1080b00e6