Attribute nve_src_intf causing bad behavior
This commit recognizes nve_src_intf in the list of Nexus Switch
config attributes as not being that of a host name.  This anomaly
came in when implementing scaling changes. Specifically the issue
is in method _get_switch_interfaces in the review/commit
https://review.openstack.org/#/c/238206
which excludes all known 'named' attributes leaving only host name
configurations.

This commit does not include tests because there is another VXLAN
change which will follow-up after this commit.  This latter change
will be patched in more branches which is why I'm separating the
commits.

Change-Id: I0a6a3bd31e52cbc7ca8aed908371b1849e83e522
Closes-bug:  #1551827