Do not update host_id on ports for routers in hosting devices
The callback from l3 config agent to the cisco l3 router plugin
updates the host_id of the router ports to the server where the
l3 config agent runs. This is not correct as the host_id should
be that of the hosting device for the neutron router.

This patch removes the code that updates the host_id in the
callback. The host_id is instead already set by the plugging
drivers (for those plugging strategies where host_id is needed).

Closes bug: #1553100
(Cherry-picked from commit: 8d3ee67d61e434c81cb4bc4b68aa0287c92edec7)

Change-Id: Iaf0fefe96d3fe586bdb29b4cdfed8f0239db64d6