Fix for defect: Fail to clear default route
Issue: The cfg agent fails when clearing default route for the tenant
routers when the external gateway is cleared.

This happens only when the tenant router is the last tenant router
with an external gateway. Thus when it is cleared the global router
(owning the external interface) is also removed.  But currently global
routers are deleted before the tenant router is processed. As part of
the global router delete, the external subinterface is also removed.
But default route for the tenant router was based on this and when the
cfg agent comes to this part, it fails.

We have now fixed the order of this processing, so that removed routers
are processed after updated routers are processed. Also global routers
are processed last in either way.

Closes-bug: #1551781
Change-Id: If76edd8ff774c9d59da1098c0f607b2ad426fa59