Use public method execute() for client extension
The client extension was using take_action(), which is really a
private method (and caused problems recently, when it changed from
get_data to take_action).

Instead, this commit uses execute(), which is the public method for
client extension purposes. This will prevent issues from occurring,
if/when the neutronclient changes.

This needs to be (manually) backported to Liberty as well.

Change-Id: I48ef59f10e741aa95a1f75e569ff9660e3ec5819
Partial-Bug: #1554717