Account for some transients in the db so we force agent to retry
Sometimes the DB is not quite settled in and some parts may be inconsistent
when the agent comes along asking the router dicts. Catch the port db exception as
the port is either being added or deleted and rest of our infra is not quite in sync.
So in this scenario we just tell the agent that the info for this router is incomplete
with a INFO_INCOMPLETE status on the router. When the agent sees this flag in its
processing it skips processing and bumps the specific router back to the
updated_routers queue to pull again the next time.
Port commit from internal (asr1k-stable-kilo) (270ce3ca514992b48f2ed59e38d6e60d4a264be7)

Change-Id: Iaf5d00e1cdcbe42c91b106a2a068a139fa0952ae
Closes-Bug: #1552031