Address double-resync of hardware hosting-devices
In this fix, the cfg-agent will no longer inadvertently trigger a
double resync if connectivity to a hosting-device is only temporarily
lost. The server/plugin is also notified when a hosting device becomes
dead or is revived. The actual heart-beat task has also been corrected
to be more consistent. Prior to this fix, the main task thread would
dominate over the heartbeat task thread. This led to extremely late
detection of hosting device outages and revivals.

Closes-Bug: 1571920
(cherry-picked from commit: 8906294)
Change-Id: I73010f4024960311ff0c8902882289989c72558f