Add routertype driver callouts to L3 Plugin
The L3 plugin doesn't have callouts for all of the Router
Type driver APIs. This patch adds the missing callouts. It
also adds a create floating IP postcommit API to the Router
Type driver APIs, adds this API to the existing drivers, and
adds the callout for it in the L3 plugin.

Closes-Bug: #1603269

Change-Id: I81277eaf0e0cfbfb317d33d801cd7341c887c73d
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>