Moved deprecated ATTR_NOT_SPECIFIED to neutron_lib
The constant ATTR_NOT_SPECIFIED and verifier is_attr_set have
been deprecated, however the new version found in neutron_lib
is incompatible with older versions of neutron.

This fix adds a version check to determine which version
should be used, and as such handles the deprecation moving forward.
Please feel free to suggest a better name for the
added file and imports.

Change-Id: Ia6c27b19b14371633f89b3c4df0f7fde515a871a