[apic-driver] avoid internal objects synchronization
In the APIC driver, users can create a network with
name 'apic-sync-network' to trigger backend full
synchronization.
During that process, the synchronizer might try to
synchronize the sync-network itself resulting in a
error and, even worse, a further resync. Avoid this
problem by discarding synchronization requests for
objects intended for internal use.

Change-Id: I537f825719ea0367e44e22cd1ce88c56b2e368b1
Closes-Bug: 1623622