Fixes agent scheduler related UT failures
From Mitaka to Newton releases changes were made in
Neutron in how DHCP notifications are triggered.
Another change was that ml2 was hardcoded to be the
core plugin in the UTs. Both of these changes caused
UTs to fail for one of the Cisco L3 router plugins.

This patch makes modifications so that those failing
UTs again pass for both Mitaka and Newton.

Change-Id: I732f246dd58a51f8ef05bc81a9c27782a19cd3ef
Closes-bug: #1625146