Restore coverage test operation
This has been broken for a long time. This commit focusing on changing
two things. First, the constraints are not applied, and as a result, we
get errors from neutron.exceptions not having EXTENDED_ATTRIBUTES_2_0.

Second, for local runs of the cover target, it was not correctly handling
uncommitted changes. The script was modified to fix "git stash".

There are code issues that prevent coverage from running, however, they
will be addressed under separate commits.

Change-Id: Ib9e4c04f7104baf0b92033c16f2520b6c428c6cc