ML2 Nexus driver fails on SessionCloseError
The ML2 Nexus ncclient(ssh) driver fails on connect call to Nexus 9K.
Failure seen in logs is as follows:
  ncclient.transport.errors.SessionCloseError: Unexpected session close
This error is seen with more recent images of the N9K when handling
yang data was introduced.  As a result, our driver must specify
that we plan to use ssh with XML data.  This is done by adding
'ssh_subsystem_name': 'xmlagent' to device_params. This change was
also verified against a Nexus3K which had a very old image.

Change-Id: I0ae3e3f51cd3bec4647fd24a9715ae42f46ee0a6
Closes-bug:  #1664293