Disables return code check in REST API update calls from some unit tests
As a temporary workaround for bug #1676435 this patch disables the return
code check in the REST API update call for some ASR1k router type driver
unit tests. This will reduce the number of Jenkins failures while the
root caused of the unit test failures are investigated.

AKA: DE3463

Partial-Bug: #1676435

Change-Id: Ia411e2a9a3fce9fee9ed598357aef377c504592e