Update tox_install.sh to align for sphinx jobs
The updates to the sphinx docs jobs in support of the updates to
the PTI wound up exposing an unintended interface. There are two flavors
of the tox_install.sh file out there, and we basically need to collapse
them into one flavor.

This script is more complex than others since it checks out different
branches. Let's leave it as is - just fix the external API to match
other projects and pass constraints file as first argument.

Change-Id: I7ca5f1df47b6b9f44297c3f0270740c99c5da18d