Fix bug with RemainderOpt not taking account of deprecations
The RemainderOpt config option type is designed to eat the remainder of
the options defined in a config file inside a config group. To ensure it
only eats the remainder of the options it pulls a list of all the other
options that should be in that group so it can exclude them. The patch
makes sure it also excludes any deprecated names too so that they aren't
accidently picked up by the RemainderOpt option.

Change-Id: I1e06e4bd8a38af5be640c56f38214bf102ca8a94