Change to new UCSM SDK
The UcsSdk has now been deprecated so change the ucsm ml2 plugin to use
the new ucsmsdk by default. The plugin will continue to work with the
UcsSdk for a period of time if the ucsmsdk has not been installed but no
new features will be available if you choose not to upgrade.

Change-Id: I9ce511157ee2a04f0601d3dc7116ae683a8f5d34