Fix readthedocs documentation building due to install from git
This patch fixes an issue causes by installing neutron from git when
building the docs on readthedocs.org. When it installs from git it runs
setup.py egg_info which causes a failure due to an incompatiblity
between setuptools 37.0.0, pbr 3.1.1, and sphinx 1.6.5. When installing
from a prebuilt wheel this doesn't occur as all the info is built into
the wheel. Neutron 11.0.2 is the latest release of neutron from the
stable/pike series.

Change-Id: I1653cd0b75295a7cab1b8a1a93c53fe7ff1a1f4f