Reword the Releases and Version support section of the docs
This patch rewords the releases information in the networking-cisco docs
and README to encorage people to use the latest version of
networking-cisco for their OpenStack release. The original documentation
was confusing as it listed multiple networking-cisco versions and
multiple OpenStack versions so it wasn't obvious the latest should be
used.

Change-Id: Ifc6d58d4146abb115c48f05bc868656aa05d3b96