No exception in update_postcommit & no port found
While running Redhat Certification tests, both UCSM and Nexus are
seeing a spurious update_postcommit at the same time the port_binding
is being removed. Nexus raises an exception while UCSM does not. Since
this update seems to be benign, it is best to ignore when the port
binding is not found but report as a warning.

Change-Id: Id106f126419a54781feabdef660fb49f8fafbbfa
Closes-bug:  #1758368