Add cisco providernet extension to Nexus doc
The existing documentation is missing mention of the extension driver
'cisco_providernet_ext'.  This driver is responsible for identifying
when a vlanid is that of a provider.  Configuration options
'provider_vlan_auto_create' and 'provider_vlan_auto_trunk'
depend on this identification to determine whether or not to
configure the vlan or apply vlan to an interface on the Nexus Switch.

Change-Id: I04436fa1284c75bcc6e3615435f656a6834c9c06
Closes-bug:  #1793146